Handling of seasonless shows, ie anime-style s1
Allow ep to match sequence_regexps, ie releases with only one number
in title. Anything caught like this is assumed to be from season one.
Parser attribute 'seasonless' is set to True if release's season has
been assumed and this is stored in episode_releases.

Emit_series will check downloaded releases for current season for
seasonless == True and search for 'season e#' on top of the usual
search strings if True.